Update cachematrix.R
Note - thanks to a bit of grading and some review in general, I found that an internally stored object can be stored and updated within the object: that I did not know.  Also, this method of coding hides the internal data members from the global environment - essentially making the data private.  I also like that.